+++ lib/flutter_fit_utils_provider.dart
@@ -1,4 +1,6 @@
 library flutter_fit_utils_provider;
 
-export 'basic_provider.dart';
-export 'form_provider.dart';
+export 'providers/fit_provider.dart';
+export 'providers/fit_form_provider.dart';
+export 'providers/data_provider.dart';
+export 'providers/items_provider.dart';